+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - use absolute imports and module path aliases
 - user: serve images at double size to maximize image clarity
 - user: reorder stars/forks per language, show stars first
+- use heroicons/react icons
 
 ## [0.7.0] - 2021-04-17
 #### Added